docs/requirements.txt
sphinxcontrib.katex
pytest
git+https://github.com/pytorch/pytorch_sphinx_theme.git
